requirements.txt
transformers==4.40.0
datasets>=2.18.0
accelerate>=0.27.0
scikit-surprise==1.1.3
pandas
numpy
tqdm
